requirements.txt
pymupdf
stix2
httpx
python-dotenv
ioc-finder
validators
tldextract
